fix bug with filter calc

+++ rs_src/gamma.rs
@@ -1,7 +1,7 @@
-use std::f64::consts::{E, PI};
+use std::f32::consts::{E, PI};
 
-const GAMMA_R: f64 = 10.900511;
-const GAMMA_DK: &[f64] = &[
+const GAMMA_R: f32 = 10.900511;
+const GAMMA_DK: &[f32] = &[
     2.48574089138753565546e-5,
     1.05142378581721974210,
     -3.45687097222016235469,
@@ -14,20 +14,20 @@ const GAMMA_DK: &[f64] = &[
     4.63399473359905636708e-6,
     -2.71994908488607703910e-9,
 ];
-const TWO_SQRT_E_OVER_PI: f64 = 1.8603827342052657173362492472666631120594218414085755;
+const TWO_SQRT_E_OVER_PI: f32 = 1.8603827342052657173362492472666631120594218414085755;
 
 /// Computes the gamma function with an accuracy
 /// of 16 floating point digits. The implementation
 /// is derived from "An Analysis of the Lanczos Gamma Approximation",
 /// Glendon Ralph Pugh, 2004 p. 116
 /// https://docs.rs/statrs/latest/src/statrs/function/gamma.rs.html#64-86
-pub fn gamma(x: f64) -> f64 {
+pub fn gamma(x: f32) -> f32 {
   if x < 0.5 {
       let s = GAMMA_DK
           .iter()
           .enumerate()
           .skip(1)
-          .fold(GAMMA_DK[0], |s, t| s + t.1 / (t.0 as f64 - x));
+          .fold(GAMMA_DK[0], |s, t| s + t.1 / (t.0 as f32 - x));
 
       PI
           / ((PI * x).sin()
@@ -39,7 +39,7 @@ pub fn gamma(x: f64) -> f64 {
           .iter()
           .enumerate()
           .skip(1)
-          .fold(GAMMA_DK[0], |s, t| s + t.1 / (x + t.0 as f64 - 1.0));
+          .fold(GAMMA_DK[0], |s, t| s + t.1 / (x + t.0 as f32 - 1.0));
 
       s * TWO_SQRT_E_OVER_PI * ((x - 0.5 + GAMMA_R) / E).powf(x - 0.5)
   }
